+++ src/methods/toArray.js
@@ -1,30 +1,28 @@
 'use strict';
 
-
 module.exports = function toArray() {
   const collectionInstance = this.constructor;
 
   const iterate = function (list, collection) {
-    list.forEach((element) => {
-      if (element instanceof collectionInstance) {
-        collection.push(element.items);
-      } else if (Array.isArray(element)) {
-        if (element.filter(el => el instanceof collectionInstance).length) {
-          iterate(element, collection);
-        } else {
-          collection.push(element);
-        }
-      }
+    const childCollection = [];
 
-
-    });
+    if (list instanceof collectionInstance) {
+      list.items.forEach(i => iterate(i, childCollection));
+      collection.push(childCollection);
+    } else if (Array.isArray(list)) {
+      list.forEach(i => iterate(i, childCollection));
+      collection.push(childCollection);
+    } else {
+      collection.push(list);
+    }
   };
 
   if (Array.isArray(this.items)) {
     const collection = [];
-    iterate(this.items, collection);
 
-    console.log(collection);
+    this.items.forEach((items) => {
+      iterate(items, collection);
+    });
 
     return collection;
   }